Resolução de bugs e adição de varias informações e novas funcionalidades como reserva(não funcional), dashboard, lista de pousadas, pesquisa, trazer todas as reservas, adição de quartos

+++ reserva-pousadas-back-end/src/main/java/com/pousadas/reserva_pousadas/controller/QuartosController.java
@@ -29,7 +29,7 @@ public class QuartosController {
     }
 
     @GetMapping("/{id}")
-    public Quartos buscarIdQuartos(@PathVariable Integer id) {
+    public Quartos buscarIdQuartos(@PathVariable Long id) {
         return quartosService.buscarPorIdQuartos(id);
     }
 
@@ -39,7 +39,7 @@ public class QuartosController {
     }
 
     @DeleteMapping("/{id}")
-    public void deletarQuarto(@PathVariable Integer id) {
+    public void deletarQuarto(@PathVariable Long id) {
         quartosService.deletarQuartos(id);
     }
