+++ reserva-pousadas-back-end/src/main/java/com/pousadas/reserva_pousadas/repository/QuartosRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pousadas.reserva_pousadas.model.Quartos;
 
-public interface QuartosRepo extends JpaRepository<Quartos, Integer> {
+public interface QuartosRepo extends JpaRepository<Quartos, Long> {
 
 }
