+++ reserva-pousadas-back-end/src/main/java/com/pousadas/reserva_pousadas/repository/ReservaRepo.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.pousadas.reserva_pousadas.model.Reserva;
 
-public interface ReservaRepo extends JpaRepository<Reserva, Integer> {
+public interface ReservaRepo extends JpaRepository<Reserva, Long> {
 
     @Query("SELECT CASE WHEN COUNT(r) > 0 THEN true ELSE false END " +
            "FROM Reserva r " +
